refactor(frontend): rename axios instance module
